requirements.txt
oslo.config
pbr
Jinja2
